+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Support raw audio files, which are essentially Wav files without their format
   header (ie just the samples).
+- Expose some of the internal utils, which may be useful for other serialization
+  tasks (eg BytesReader and BytesWriter).
 
 ## 1.2.0